requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
azure-core==1.35.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
beanie==1.29.0
beautifulsoup4==4.13.4
billiard==4.2.1
Brotli==1.1.0
cachetools==5.5.2
camelot-py==0.11.0
celery==5.5.2
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cssselect2==0.8.0
curl_cffi==0.11.4
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==8.0.1
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.5
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.5.1
ghostscript==0.8.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth-oauthlib==1.2.2
google-auth==2.40.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio-status==1.73.1
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx-sse==0.4.1
httpx==0.28.1
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipinfo==5.1.1
ipython==8.36.0
isodate==0.7.2
janus==2.0.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-specifications==2025.4.1
jsonschema==4.24.0
kiwisolver==1.4.8
kombu==5.5.4
langchain-community==0.3.23
langchain-core==0.3.72
langchain-google-genai==2.1.4
langchain-openai==0.3.16
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langchain==0.3.27
langgraph-checkpoint==2.0.26    
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langgraph==0.4.1
langsmith==0.3.45
lazy-model==0.2.0
litellm==1.61.0
lxml==5.4.0
markdown-it-py==3.0.0
Markdown==3.6
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
matplotlib==3.10.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
motor==3.7.0
multidict==6.6.3
multitasking==0.0.11
mypy_extensions==1.1.0
mypy==1.16.1
narwhals==1.45.0
numpy==2.2.5
oauthlib==3.3.1
openai==1.75.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.1
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.1
pdfkit==1.0.0
pdfminer.six==20250506
pdfplumber==0.11.7
pdftopng==0.2.4
peewee==3.18.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
portalocker==2.10.1
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1_modules==0.4.1
pyasn1==0.4.8
pycparser==2.22
pydantic_core==2.33.2
pydantic-settings==2.10.1
pydantic==2.11.7
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.12.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests-file==2.1.0
requests-futures==1.0.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
selenium==4.32.0
six==1.17.0
smolagents==1.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlalchemy==2.0.42
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
tabulate==0.9.0
tavily-python==0.7.10
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
trio-websocket==0.12.2
trio==0.30.0
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yahooquery==2.4.1
yarl==1.20.1
yfinance==0.2.61
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0